update language and refactor main statistic
